[build] Move Isa enum to the meta library;
