+++ lib/codegen/meta/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod error;
 pub mod gen_types;
+pub mod isa;
 
 mod base;
 mod cdsl;
